+++ sdks/java/core/src/test/java/org/apache/beam/sdk/transforms/reflect/DoFnSignaturesTest.java
@@ -28,6 +28,7 @@ import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.lang.reflect.Field;
@@ -35,10 +36,13 @@ import java.util.List;
 import org.apache.beam.sdk.coders.StringUtf8Coder;
 import org.apache.beam.sdk.coders.VarIntCoder;
 import org.apache.beam.sdk.coders.VarLongCoder;
+import org.apache.beam.sdk.io.range.OffsetRange;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.schemas.FieldAccessDescriptor;
 import org.apache.beam.sdk.state.CombiningState;
 import org.apache.beam.sdk.state.GroupingState;
+import org.apache.beam.sdk.state.MapState;
+import org.apache.beam.sdk.state.SetState;
 import org.apache.beam.sdk.state.StateSpec;
 import org.apache.beam.sdk.state.StateSpecs;
 import org.apache.beam.sdk.state.TimeDomain;
@@ -47,6 +51,7 @@ import org.apache.beam.sdk.state.TimerSpec;
 import org.apache.beam.sdk.state.TimerSpecs;
 import org.apache.beam.sdk.state.ValueState;
 import org.apache.beam.sdk.state.WatermarkHoldState;
+import org.apache.beam.sdk.testing.SerializableMatchers;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.Sum;
 import org.apache.beam.sdk.transforms.reflect.DoFnSignature.Parameter;
@@ -64,12 +69,15 @@ import org.apache.beam.sdk.transforms.reflect.DoFnSignature.Parameter.TimerParam
 import org.apache.beam.sdk.transforms.reflect.DoFnSignature.Parameter.TimestampParameter;
 import org.apache.beam.sdk.transforms.reflect.DoFnSignature.Parameter.WindowParameter;
 import org.apache.beam.sdk.transforms.reflect.DoFnSignaturesTestUtils.FakeDoFn;
+import org.apache.beam.sdk.transforms.splittabledofn.RestrictionTracker;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
 import org.apache.beam.sdk.transforms.windowing.PaneInfo;
+import org.apache.beam.sdk.transforms.windowing.TimestampCombiner;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.Row;
 import org.apache.beam.sdk.values.TypeDescriptor;
 import org.apache.beam.sdk.values.TypeDescriptors;
+import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.Lists;
 import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;
 import org.joda.time.Instant;
@@ -830,6 +838,43 @@ public class DoFnSignaturesTest {
   }
 
   @Test
+  public void testStateParameterAlwaysFetched() {
+    thrown.expect(IllegalArgumentException.class);
+    thrown.expectMessage("ReadableStates");
+    DoFnSignature sig =
+        DoFnSignatures.getSignature(
+            new DoFn<KV<String, Integer>, Long>() {
+              @StateId("my-id")
+              private final StateSpec<MapState<Integer, Integer>> myfield =
+                  StateSpecs.map(VarIntCoder.of(), VarIntCoder.of());
+
+              @ProcessElement
+              public void myProcessElement(
+                  ProcessContext context,
+                  @AlwaysFetched @StateId("my-id") MapState<Integer, Integer> one) {}
+            }.getClass());
+    StateParameter stateParameter = (StateParameter) sig.processElement().extraParameters().get(1);
+    assertTrue(stateParameter.alwaysFetched());
+  }
+
+  @Test
+  public void testStateParameterAlwaysFetchNonReadableState() {
+    thrown.expect(IllegalArgumentException.class);
+    thrown.expectMessage("ReadableStates");
+    DoFnSignatures.getSignature(
+        new DoFn<KV<String, Integer>, Long>() {
+          @StateId("my-id")
+          private final StateSpec<MapState<Integer, Integer>> myfield =
+              StateSpecs.map(VarIntCoder.of(), VarIntCoder.of());
+
+          @ProcessElement
+          public void myProcessElement(
+              ProcessContext context,
+              @AlwaysFetched @StateId("my-id") MapState<Integer, Integer> one) {}
+        }.getClass());
+  }
+
+  @Test
   public void testStateParameterDuplicate() throws Exception {
     thrown.expect(IllegalArgumentException.class);
     thrown.expectMessage("duplicate");
@@ -980,7 +1025,7 @@ public class DoFnSignaturesTest {
     DoFnSignature.StateDeclaration decl =
         sig.stateDeclarations().get(DoFnOverridingAbstractStateUse.STATE_ID);
     StateParameter stateParam = (StateParameter) sig.processElement().extraParameters().get(1);
-
+    assertFalse(stateParam.alwaysFetched());
     assertThat(
         decl.field(),
         equalTo(DoFnDeclaringStateAndAbstractUse.class.getDeclaredField("myStateSpec")));
@@ -1190,6 +1235,239 @@ public class DoFnSignaturesTest {
     assertThat(params.get(2), instanceOf(PipelineOptionsParameter.class));
     assertThat(params.get(3), instanceOf(OutputReceiverParameter.class));
     assertThat(params.get(4), instanceOf(TaggedOutputReceiverParameter.class));
+  }
+
+  private interface FeatureTest {
+    void test();
+  }
+
+  private static class StatelessDoFn extends DoFn<String, String> implements FeatureTest {
+    @ProcessElement
+    public void process(@Element String input) {}
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(false));
+    }
+  }
+
+  private static class StatefulWithValueState extends DoFn<KV<String, String>, String>
+      implements FeatureTest {
+    @StateId("state")
+    private final StateSpec<ValueState<String>> state = StateSpecs.value();
+
+    @ProcessElement
+    public void process(@Element KV<String, String> input) {}
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(false));
+    }
+  }
+
+  private static class StatefulWithTimers extends DoFn<KV<String, String>, String>
+      implements FeatureTest {
+    @TimerId("timer")
+    private final TimerSpec spec = TimerSpecs.timer(TimeDomain.EVENT_TIME);
+
+    @ProcessElement
+    public void process(@Element KV<String, String> input) {}
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(false));
+    }
+
+    @OnTimer("timer")
+    public void onTimer() {}
+  }
+
+  private static class StatefulWithTimersAndValueState extends DoFn<KV<String, String>, String>
+      implements FeatureTest {
+    @TimerId("timer")
+    private final TimerSpec timer = TimerSpecs.timer(TimeDomain.EVENT_TIME);
+
+    @StateId("state")
+    private final StateSpec<SetState<String>> state = StateSpecs.set();
+
+    @ProcessElement
+    public void process(@Element KV<String, String> input) {}
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(false));
+    }
+
+    @OnTimer("timer")
+    public void onTimer() {}
+  }
+
+  private static class StatefulWithSetState extends DoFn<KV<String, String>, String>
+      implements FeatureTest {
+    @StateId("state")
+    private final StateSpec<SetState<String>> spec = StateSpecs.set();
+
+    @ProcessElement
+    public void process(@Element KV<String, String> input) {}
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(false));
+    }
+  }
+
+  private static class StatefulWithMapState extends DoFn<KV<String, String>, String>
+      implements FeatureTest {
+    @StateId("state")
+    private final StateSpec<MapState<String, String>> spec = StateSpecs.map();
+
+    @ProcessElement
+    public void process(@Element KV<String, String> input) {}
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(false));
+    }
+  }
+
+  private static class StatefulWithWatermarkHoldState extends DoFn<KV<String, String>, String>
+      implements FeatureTest {
+    @StateId("state")
+    private final StateSpec<WatermarkHoldState> spec =
+        StateSpecs.watermarkStateInternal(TimestampCombiner.LATEST);
+
+    @ProcessElement
+    public void process(@Element KV<String, String> input) {}
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(false));
+    }
+  }
+
+  private static class RequiresTimeSortedInput extends DoFn<KV<String, String>, String>
+      implements FeatureTest {
+    @ProcessElement
+    @RequiresTimeSortedInput
+    public void process(@Element KV<String, String> input) {}
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(true));
+    }
+  }
+
+  private static class Splittable extends DoFn<KV<String, Long>, String> implements FeatureTest {
+    @ProcessElement
+    public void process(ProcessContext c, RestrictionTracker<OffsetRange, ?> tracker) {}
+
+    @GetInitialRestriction
+    public OffsetRange getInitialRange(@Element KV<String, Long> element) {
+      return new OffsetRange(0L, element.getValue());
+    }
+
+    @Override
+    public void test() {
+      assertThat(DoFnSignatures.isSplittable(this), SerializableMatchers.equalTo(true));
+      assertThat(DoFnSignatures.isStateful(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesTimers(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesBagState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesMapState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesSetState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesValueState(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.usesWatermarkHold(this), SerializableMatchers.equalTo(false));
+      assertThat(DoFnSignatures.requiresTimeSortedInput(this), SerializableMatchers.equalTo(false));
+    }
+  }
+
+  private final List<FeatureTest> tests =
+      Lists.newArrayList(
+          new StatelessDoFn(),
+          new StatefulWithValueState(),
+          new StatefulWithTimers(),
+          new StatefulWithTimersAndValueState(),
+          new StatefulWithSetState(),
+          new StatefulWithMapState(),
+          new StatefulWithWatermarkHoldState(),
+          new RequiresTimeSortedInput(),
+          new Splittable());
+
+  @Test
+  public void testAllDoFnFeatures() {
+    tests.forEach(FeatureTest::test);
   }
 
   private Matcher<String> mentionsTimers() {