+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/spanner/NaiveSpannerRead.java
@@ -81,7 +81,7 @@ abstract class NaiveSpannerRead
 
     @Setup
     public void setup() throws Exception {
-      spannerAccessor = config.connectToSpanner();
+      spannerAccessor = SpannerAccessor.create(config);
     }
 
     @Teardown