+++ sdks/java/io/hadoop-file-system/src/main/java/org/apache/beam/sdk/io/hdfs/HadoopFileSystemOptions.java
@@ -28,6 +28,7 @@ import org.apache.beam.sdk.options.DefaultValueFactory;
 import org.apache.beam.sdk.options.Description;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.annotations.VisibleForTesting;
+import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Splitter;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Strings;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.Lists;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.Sets;
@@ -87,7 +88,29 @@ public interface HadoopFileSystemOptions extends PipelineOptions {
         }
       }
 
+      /*
+       * Explode the paths by ":" to handle the case in which the environment variables
+       * contains multiple paths.
+       *
+       * This happens on Cloudera 6.x, in which the spark-env.sh script sets
+       * HADOOP_CONF_DIR by appending also the Hive configuration folder:
+       *
+       * if [ -d "$HIVE_CONF_DIR" ]; then
+       *   HADOOP_CONF_DIR="$HADOOP_CONF_DIR:$HIVE_CONF_DIR"
+       * fi
+       * export HADOOP_CONF_DIR
+       *
+       */
+      Set<String> explodedConfDirs = Sets.newHashSet();
       for (String confDir : confDirs) {
+        Iterable<String> paths = Splitter.on(':').split(confDir);
+        for (String p : paths) {
+          explodedConfDirs.add(p);
+        }
+      }
+
+      // Load the configuration from paths found (if exists)
+      for (String confDir : explodedConfDirs) {
         if (new File(confDir).exists()) {
           Configuration conf = new Configuration(false);
           boolean confLoaded = false;